Fix cognitive complexity
